Move v130 gen xml tests
